добавление activity_media.xml, activity_search.xml; переходы в actyvity_settings.xml, activity_media.xml, activity_search.xml
