+++ app/src/main/java/com/ealyk/playlistmaker2/SettingsActivity.kt
@@ -1,0 +1,51 @@
+package com.ealyk.playlistmaker2
+
+import android.annotation.SuppressLint
+import android.content.Intent
+import android.content.SharedPreferences
+import android.os.Bundle
+import android.os.PersistableBundle
+import android.widget.Button
+import android.widget.Switch
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatDelegate
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+
+class SettingsActivity : AppCompatActivity() {
+    @SuppressLint("MissingInflatedId")
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_settings)
+
+        val themeSwitch = findViewById<Switch>(R.id.switch_theme)
+        val sharedPreferences: SharedPreferences = getSharedPreferences("themePrefs", MODE_PRIVATE)
+        val editor = sharedPreferences.edit()
+        val isNightMode = sharedPreferences.getBoolean("NightMode", false)
+        themeSwitch.isChecked = isNightMode
+        if (isNightMode){
+            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
+        }
+        else {
+            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
+        }
+        themeSwitch.setOnCheckedChangeListener { _, isChecked ->
+            if (isChecked){
+                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
+                editor.putBoolean("NightMode", true)
+            }
+            else {
+                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
+                editor.putBoolean("NightMode", false)
+            }
+            editor.apply()
+        }
+        val backButton = findViewById<Button>(R.id.back_button)
+        backButton.setOnClickListener {
+            val backButtonIntend = Intent(this, MainActivity::class.java)
+            startActivity(backButtonIntend)
+        }
+    }
+
+}
