+++ app/src/main/res/layout/activity_settings.xml
@@ -3,6 +3,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    tools:context=".SettingsActivity"
+    android:id="@+id/settings"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/background_settings">
     <LinearLayout
@@ -21,6 +23,7 @@
 >
             <Button
                 android:layout_width="@dimen/header_buttons_weight"
+                android:id="@+id/back_button"
                 android:layout_height="wrap_content"
                 android:textColor="@android:color/transparent"
                 tools:ignore="TouchTargetSizeCheck"
